+++ next.config.ts
@@ -1,5 +1,6 @@
 import type { NextConfig } from "next";
 import { defaultConfig } from "next/dist/server/config-shared";
+import pkg from "./package.json";
 
 const cspHeader = `
   script-src 'self' 'unsafe-eval' 'unsafe-inline';
@@ -12,32 +13,68 @@ const cspHeader = `
   frame-ancestors 'none';
 `;
 
+const baseSecurityHeader = [
+  {
+    key: "Referrer-Policy",
+    value: "no-referrer",
+  },
+  {
+    key: "Cross-Origin-Embedder-Policy",
+    value: "require-corp",
+  },
+  {
+    key: "Cross-Origin-Opener-Policy",
+    value: "same-origin",
+  },
+  {
+    key: "Cross-Origin-Resource-Policy",
+    value: "same-origin",
+  },
+  {
+    key: "Permissions-Policy",
+    value:
+      "accelerometer=(),autoplay=(),camera=(),display-capture=(),encrypted-media=(),fullscreen=(),geolocation=(),gyroscope=(),magnetometer=(),microphone=(),midi=(),payment=(),picture-in-picture=(),publickey-credentials-get=(),screen-wake-lock=(),sync-xhr=(self),usb=(),web-share=(),xr-spatial-tracking=()",
+  },
+  {
+    key: "Strict-Transport-Security",
+    value: "max-age=31536000; includeSubDomains; preload",
+  },
+  {
+    key: "X-Content-Type-Options",
+    value: "nosniff",
+  },
+  {
+    key: "X-Frame-Options",
+    value: "DENY",
+  },
+  {
+    key: "X-DNS-Prefetch-Control",
+    value: "on",
+  },
+  {
+    key: "X-Permitted-Cross-Domain-Policies",
+    value: "none",
+  },
+  {
+    key: "x-powered-by",
+    value: "owlsome-official/next-tailwind-ts",
+  },
+];
+
 const nextConfig: NextConfig = {
   async headers() {
     return [
       {
         source: "/(.*)",
         headers: [
+          ...baseSecurityHeader,
           {
             key: "Content-Security-Policy",
             value: cspHeader.replace(/\n/g, ""),
           },
           {
-            key: "X-Content-Type-Options",
-            value: "nosniff",
-          },
-          {
-            key: "X-Frame-Options",
-            value: "SAMEORIGIN",
-          },
-        ],
-      },
-      {
-        source: "/support",
-        headers: [
-          {
-            key: "x-powered-by",
-            value: "owlsome-official/next-tailwind-ts",
+            key: "X-App-Version",
+            value: pkg.version,
           },
         ],
       },